unify sqlite models
